requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.5
django-crispy-forms==2.1
django-jazzmin==3.0.0
greenlet==3.0.3
msgpack==1.0.8
neovim==0.3.1
pillow==10.3.0
pynvim==0.5.0
sqlparse==0.5.0
tornado==6.4
typing_extensions==4.11.0
